backendManim/requirements.txt
fastapi
uvicorn
manim
google-generativeai
python-dotenv
pydantic
pydantic-settings
celery[redis]  # NEW
redis  # NEW